apps/ingestion/streaming/replay_producer/requirements.txt
fastavro==1.12.1
kafka-python==2.3.1
pyarrow==23.0.1
